Fixed some lint errors and moved the interfaces on shared
