+++ src/app/features/admin/components/users/user-list/user-list.component.html
@@ -1,1 +1,3 @@
-<p>user-list works!</p>
+<div class="users-view-container dark:text-light-text-tertiary">
+<p>test</p>
+</div>
