+++ src/app/features/admin/components/users/user-list/user-list.component.scss
@@ -1,0 +1,3 @@
+.users-view-container {
+  padding: 1rem;
+}
