+++ src/app/features/articles/services/article.service.ts
@@ -5,7 +5,7 @@ import {
   Article,
   FilterArticles,
   PaginatedResponse,
-} from '../../../core/interfaces/core.interfaces';
+} from '../../../shared/interfaces/core.interfaces';
 import { Subscription } from 'rxjs';
 import { Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
