+++ src/app/shared/interfaces/core.interfaces.ts
