+++ src/app/tests/articles/article.service.spec.ts
@@ -5,7 +5,7 @@ import {
 } from '@angular/common/http/testing';
 import { ArticleService } from '../../features/articles/services/article.service';
 import { HttpErrorResponse, provideHttpClient } from '@angular/common/http';
-import { Article } from '../../core/interfaces/core.interfaces';
+import { Article } from '../../shared/interfaces/core.interfaces';
 import { ErrorService } from '../../shared/services/error.service';
 
 describe('ArticleService', () => {
